Use pimpl idiom in fmm_evaluator/operator.
